certloader: Fix reload on K8s Secret/ConfigMap update
This change fixes the certloader `Watch` function to support the
fsnotify sequence emitted when Kubernetes updates a volume mount for a
ConfigMap or Secret.

The previous code watched for changes in files by adding watchers
directly on the files themselves, or their parent directory if the files
were absent. This works fine unless Kubernetes is used to mount the
keyfiles or certificates via ConfigMap or Secret. In such a case, the
projected files are symlinks containg a double indirection.

This means that when the file was derived from a Kubernetes ConfigMap or
Secret, the previous code would watch the target of the symlink to the
fsnotify watcher. This would cause it to only receive an event the first
time the files were updated, as the target file would be deleted
afterwards (and therefore no new events would be produced for that
watched file).

Due to the double indirection, it is not sufficient to watch the symlink
either, see
https://github.com/spf13/viper/pull/415#issuecomment-344050279 for an
example of the fsnotify events fired when Kubernetes updates the files.

Therefore, instead of trying to deal with all Kubernetes corner-cases,
this change implements a much simpler and hopefully more robust
approach.

If the code detects that a path is managed by a Kubernetes volume mount
(see findKubernetesVolumeMount in newFsWatcher), it watches the volume
mount directory (non-recursively) instead of the file. If the underlying
volume mount is updated, we coascale all fsnotify events within a
certain time window and then reload _all_ certificates and keypairs at
once. This drastically simplifies the implementation and allows us to
ignore the fact that the files might or might not be symlinked.

As an added bonus, coascaling also deals with the non-Kubernetes case
where the keypair and ca-certificates files are updated non-atomically
in individual writes.

A unit test is added which simulates the behavior observed by
Kubernetes. If no Kubernetes volume mount is detected, we watch the
files parent directory to pick up delete + create sequences.

Fixes: #13621

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>