daemon: Join cluster before starting IPAM
Split the --jopin-cluster feature init into two phases:

1. Get config from clustermesh-apiserver by resetting node
registration to only contain the node name. clustermesh-apiserver
updates the key with configuration for this node derived from the
cluster state.

2. Update node registration after node IPAM is started with full node
information.

When node registration is started, the local cluster name may not be
known yet. To accomodate this only the node name is used to identify
the node registration.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>